fintech-crm-development/backend/requirements.txt
aiokafka==0.8.1
alembic==1.11.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
azure-common==1.1.28
azure-core==1.26.4
azure-mgmt-core==1.4.0
azure-storage-blob==12.16.0
bcrypt==4.0.1
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confluent-kafka==2.1.1
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
databases==0.7.0
dnspython==2.3.0
email-validator==2.0.0.post2
emails==0.6
fastapi==0.95.2
google-api-core==2.11.1
google-auth==2.21.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
isodate==0.6.1
kafka-python==2.0.2
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
msal==1.22.0
msal-extensions==1.0.0
numpy>=1.21.5,<1.27.0
opencensus==0.11.2
opencensus-context==0.1.3
opencensus-ext-azure==1.1.9
openpyxl>=3.0.0,<4.0.0
packaging==23.1
pandas>=1.3.0,<2.0.0
passlib==1.7.4
portalocker==2.7.0
premailer==3.10.0
protobuf==4.23.4
psutil==5.9.5
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
redis==5.0.1
aioredis==2.0.1
httpx==0.24.1
jinja2==3.1.2
reportlab
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
sqlalchemy2-stubs==0.0.2a34
starlette==0.27.0
tenacity==8.2.2
typing_extensions==4.5.0
urllib3==1.26.16
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
